requirements/prod.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.8,<1.9

# Templates
Markdown

# Logging
raven

# Celery: Message queue
celery
django-celery
django-model-utils==2.2


# form filter
django-filter==0.11.0

# tables
django-tables2==1.2.5

# Caching
python-memcached


# braces
django-braces==1.9.0

# Database backup
django-dbbackup

# Extensions
django-extensions

# Forms
django-crispy-forms
django-parsley
django-simple-captcha

# Env
django-environ==0.4.0

# Sass/Compass
django-compressor


# dj-kits
dj-kits==1.0.3
